Adding test

+++ backend/corpora/lambdas/api/v1/dataset.py
@@ -70,8 +70,11 @@ def delete_dataset(dataset_uuid: str, user: str):
     if dataset.collection_visibility == CollectionVisibility.PUBLIC:
         return make_response(jsonify("Can not delete a public dataset"), 405)
     if dataset.tombstone is False:
-        dataset.asset_deletion()
-        dataset.delete()
+        if dataset.published:
+            dataset.tombstone_dataset_and_delete_child_objects()
+        else:
+            dataset.asset_deletion()
+            dataset.delete()
     return "", 202
